CDP-1998 - export getGraphicImgsBySocial from utils, update GraphicProject, GraphicCard & test components
