+++ components/GraphicProject/GraphicProject.js
@@ -6,7 +6,10 @@ import { updateUrl } from 'lib/browser';
 import { displayDOSLogo } from 'lib/sourceLogoUtils';
 import { getCount, getFileExt, getPreviewNotificationStyles } from 'lib/utils';
 import { useAuth } from 'context/authContext';
-import { normalizeGraphicProjectByAPI } from './utils';
+import {
+  normalizeGraphicProjectByAPI,
+  getGraphicImgsBySocial
+} from './utils';
 
 import downloadIcon from 'static/icons/icon_download.svg';
 import shareIcon from 'static/icons/icon_share.svg';
@@ -59,28 +62,7 @@ const GraphicProject = ( {
   } = normalizedGraphicData;
 
   // Use Twitter graphics as default for display otherwise whatever graphic image is available
-  const getGraphicImgsBySocial = platform => {
-    if ( !getCount( images ) ) return [];
-
-    const filteredImgs = images.reduce( ( allImgs, img ) => {
-      let platformImgs = [];
-      let condition = img.social === platform;
-
-      if ( useGraphQl ) {
-        platformImgs = img.social.filter( s => s.name === platform );
-        condition = getCount( platformImgs );
-      }
-
-      if ( condition ) {
-        allImgs.push( img );
-      }
-
-      return allImgs;
-    }, [] );
-
-    return getCount( filteredImgs ) ? filteredImgs : images;
-  };
-  const graphicUnits = getGraphicImgsBySocial( 'Twitter' );
+  const graphicUnits = getGraphicImgsBySocial( images, 'Twitter', useGraphQl );
 
   // Set default unit to English lang version if available
   // unless path is for specific lang or no english version
